python/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backoff==2.2.1
certifi==2022.9.24
chardet==4.0.0
charset-normalizer==2.1.1
frozenlist==1.3.3
gql==3.4.1
graphql-core==3.2.3
idna==2.10
multidict==6.0.2
numpy==1.25.2
oauthlib==3.2.0
omnikeeper-client==3.2.0
pandas==2.1.0
python-dateutil==2.8.2
python-json-logger==2.0.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
six==1.16.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.12
webcolors==1.3
yarl==1.8.1
